LPS-61776 move package for consistency
