+++ modules/portal-security/portal-security-sso-openid/src/main/java/com/liferay/portal/security/sso/openid/configuration/OpenIdConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.openid.module.configuration;
+package com.liferay.portal.security.sso.openid.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
 
@@ -31,7 +31,7 @@ import com.liferay.configuration.admin.ConfigurationAdmin;
  */
 @ConfigurationAdmin(category = "platform")
 @Meta.OCD(
-	id = "com.liferay.portal.security.sso.openid.module.configuration.OpenIdConfiguration",
+	id = "com.liferay.portal.security.sso.openid.configuration.OpenIdConfiguration",
 	localization = "content/Language", name = "%open.id.configuration.name"
 )
 public interface OpenIdConfiguration {
