+++ modules/portal-security/portal-security-sso-openid/src/main/java/com/liferay/portal/security/sso/openid/internal/module/configuration/definition/OpenIdCompanyServiceConfigurationPidMapping.java
@@ -15,8 +15,8 @@
 package com.liferay.portal.security.sso.openid.internal.module.configuration.definition;
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
+import com.liferay.portal.security.sso.openid.configuration.OpenIdConfiguration;
 import com.liferay.portal.security.sso.openid.constants.OpenIdConstants;
-import com.liferay.portal.security.sso.openid.module.configuration.OpenIdConfiguration;
 
 import org.osgi.service.component.annotations.Component;
