+++ modules/portal-security/portal-security-sso-openid/src/main/java/com/liferay/portal/security/sso/openid/internal/portlet/action/OpenIdLoginMVCActionCommand.java
@@ -141,7 +141,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Jorge Ferrer
  */
 @Component(
-	configurationPid = "com.liferay.portal.security.sso.openid.module.configuration.OpenIdConfiguration",
+	configurationPid = "com.liferay.portal.security.sso.openid.configuration.OpenIdConfiguration",
 	immediate = true,
 	property = {
 		"javax.portlet.name=" + PortletKeys.FAST_LOGIN,
