+++ modules/portal-security/portal-security-sso-openid/src/main/resources/META-INF/resources/com.liferay.portal.settings.web/init.jsp
@@ -22,8 +22,8 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 <%@ page import="com.liferay.portal.kernel.module.configuration.ConfigurationFactoryUtil" %><%@
 page import="com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator" %><%@
 page import="com.liferay.portal.kernel.settings.ParameterMapSettingsLocator" %><%@
-page import="com.liferay.portal.security.sso.openid.constants.OpenIdConstants" %><%@
-page import="com.liferay.portal.security.sso.openid.module.configuration.OpenIdConfiguration" %>
+page import="com.liferay.portal.security.sso.openid.configuration.OpenIdConfiguration" %><%@
+page import="com.liferay.portal.security.sso.openid.constants.OpenIdConstants" %>
 
 <%@ page import="javax.portlet.ActionRequest" %>
